+++ packages/spruce-event-plugin/src/__tests__/behavioral/ListeningToEvents.test.ts
@@ -1,15 +1,15 @@
 import { MercuryClientFactory } from '@sprucelabs/mercury-client'
+import { MercuryClient } from '@sprucelabs/mercury-client'
 import {
-	buildEmitTargetAndPayloadSchema,
 	eventDiskUtil,
 	eventResponseUtil,
 } from '@sprucelabs/spruce-event-utils'
 import { diskUtil } from '@sprucelabs/spruce-skill-utils'
+import { MercuryFixture } from '@sprucelabs/spruce-test-fixtures'
 import { assert, test } from '@sprucelabs/test'
 import { EventFeature } from '../..'
-import { MercuryFixture } from '../../../../spruce-test-fixtures/build'
 import SpruceError from '../../errors/SpruceError'
-import { EventFeaturePlugin, MercuryClient } from '../../plugins/event.plugin'
+import { EventFeaturePlugin } from '../../plugins/event.plugin'
 import AbstractEventPluginTest from '../../tests/AbstractEventPluginTest'
 
 declare module '@sprucelabs/spruce-skill-utils/build/types/skill.types' {
@@ -19,14 +19,10 @@ declare module '@sprucelabs/spruce-skill-utils/build/types/skill.types' {
 }
 
 export default class ReceivingEventsTest extends AbstractEventPluginTest {
-	protected static async beforeAll() {
-		await super.beforeAll()
-	}
-
 	protected static async beforeEach() {
 		await super.beforeEach()
-		MercuryClientFactory.setIsTestMode(false)
 
+		MercuryClientFactory.setIsTestMode(false)
 		MercuryFixture.setShouldMixinCoreEventContractsWhenImportingLocal(true)
 
 		delete process.env.DID_BOOT_FIRED
@@ -211,7 +207,7 @@ export default class ReceivingEventsTest extends AbstractEventPluginTest {
 		const setShouldAutoRegistrationInvocations: boolean[] = []
 		const autoRegisterForOn: boolean[] = []
 
-		const { bootedSkill, events } = await this.registerSkillAndSetupListeners({
+		const { currentSkill, events } = await this.registerSkillAndSetupListeners({
 			onUnregisterListeners: () => {
 				unRegisterListenerCount++
 			},
@@ -228,7 +224,7 @@ export default class ReceivingEventsTest extends AbstractEventPluginTest {
 			},
 		})
 
-		await this.bootKillAndResetSkill(bootedSkill, events)
+		await this.bootKillAndResetSkill(currentSkill, events)
 
 		assert.isLength(
 			setShouldAutoRegistrationInvocations,
@@ -239,7 +235,7 @@ export default class ReceivingEventsTest extends AbstractEventPluginTest {
 		assert.isTrue(setShouldAutoRegistrationInvocations[1])
 		assert.isTrue(autoRegisterForOn[0])
 
-		await this.bootKillAndResetSkill(bootedSkill, events)
+		await this.bootKillAndResetSkill(currentSkill, events)
 
 		assert.isLength(setShouldAutoRegistrationInvocations, 4)
 		assert.isFalse(setShouldAutoRegistrationInvocations[2])
@@ -258,7 +254,7 @@ export default class ReceivingEventsTest extends AbstractEventPluginTest {
 
 		diskUtil.writeFile(listenerDest, 'exports.default = function() {}')
 
-		await this.bootKillAndResetSkill(bootedSkill, events)
+		await this.bootKillAndResetSkill(currentSkill, events)
 
 		assert.isLength(setShouldAutoRegistrationInvocations, 6)
 		assert.isTrue(setShouldAutoRegistrationInvocations[4])
@@ -278,7 +274,7 @@ export default class ReceivingEventsTest extends AbstractEventPluginTest {
 		let unRegisterListenerCount = 0
 		const shoulds: boolean[] = []
 
-		const { bootedSkill, events } = await this.registerSkillAndSetupListeners({
+		const { currentSkill, events } = await this.registerSkillAndSetupListeners({
 			onUnregisterListeners: () => {
 				unRegisterListenerCount++
 			},
@@ -289,8 +285,8 @@ export default class ReceivingEventsTest extends AbstractEventPluginTest {
 			onAttachListener: () => {},
 		})
 
-		await this.bootKillAndResetSkill(bootedSkill, events)
-		await this.bootKillAndResetSkill(bootedSkill, events)
+		await this.bootKillAndResetSkill(currentSkill, events)
+		await this.bootKillAndResetSkill(currentSkill, events)
 
 		assert.isEqual(unRegisterListenerCount, 2)
 		assert.isLength(shoulds, 4)
@@ -304,16 +300,16 @@ export default class ReceivingEventsTest extends AbstractEventPluginTest {
 	protected static async willlReRegisterListenersWithDifferentHost() {
 		let unRegisterListenerCount = 0
 
-		const { bootedSkill, events } = await this.registerSkillAndSetupListeners({
+		const { currentSkill, events } = await this.registerSkillAndSetupListeners({
 			onUnregisterListeners: () => {
 				unRegisterListenerCount++
 			},
 		})
 
-		await this.bootKillAndResetSkill(bootedSkill, events)
+		await this.bootKillAndResetSkill(currentSkill, events)
 		process.env.HOST = process.env.HOST + ':443'
 
-		await this.bootKillAndResetSkill(bootedSkill, events)
+		await this.bootKillAndResetSkill(currentSkill, events)
 
 		assert.isEqual(unRegisterListenerCount, 2)
 	}
@@ -368,7 +364,7 @@ export default class ReceivingEventsTest extends AbstractEventPluginTest {
 	}
 
 	private static async setupTwoSkillsAndBoot(dirName: string) {
-		this.cwd = await this.setupSkillDir(dirName)
+		this.cwd = await this.generateSkillFromTestPath(dirName)
 
 		const { skill: skill1, client: client1 } = await this.Fixture(
 			'skill'
@@ -405,73 +401,18 @@ export default class ReceivingEventsTest extends AbstractEventPluginTest {
 	}
 
 	private static setupListenersForEventsRegisteredBySkill(skill: any) {
-		diskUtil.moveDir(
-			this.resolvePath('build/listeners/namespace'),
-			this.resolvePath(`build/listeners/`, skill.slug)
-		)
+		return this.EventFixture().copyListenersIntoPlace(skill.slug)
 	}
 
-	private static async setupSkillDir(dirName = 'registered-skill') {
-		const source = this.resolveTestPath(dirName)
-		const destination = this.resolveTestPath(`${new Date().getTime()}/skill`)
-
-		await diskUtil.copyDir(source, destination)
-
-		return destination
-	}
-
-	private static async registerEvents(client: any, eventName: string) {
-		const contract = this.buildContract(eventName)
-		const results = await client.emit(`register-events::v2020_12_25`, {
-			payload: {
-				contract,
-			},
-		})
-
-		eventResponseUtil.getFirstResponseOrThrow(results)
+	private static async registerEvents(
+		client: MercuryClient,
+		eventName: string
+	) {
+		return this.EventFixture().registerEvents(client, eventName)
 	}
 
 	private static buildContract(eventName: string) {
-		return {
-			eventSignatures: {
-				[eventName]: {
-					emitPayloadSchema: buildEmitTargetAndPayloadSchema({
-						eventName,
-						targetSchema: {
-							id: 'emitTarget',
-							fields: {
-								organizationId: {
-									type: 'id',
-									isRequired: true,
-								},
-							},
-						},
-						payloadSchema: {
-							id: 'emitPayload',
-							fields: {
-								foo: {
-									type: 'text',
-								},
-								bar: {
-									type: 'text',
-								},
-								orgId: {
-									type: 'text',
-								},
-							},
-						},
-					}),
-					responsePayloadSchema: {
-						id: 'responsePayload',
-						fields: {
-							taco: {
-								type: 'text',
-							},
-						},
-					},
-				},
-			},
-		}
+		return this.EventFixture().buildContract(eventName)
 	}
 
 	protected static async registerSkillAndSetupListeners(options?: {
@@ -480,71 +421,7 @@ export default class ReceivingEventsTest extends AbstractEventPluginTest {
 		onSetShouldAutoRegisterListeners?: (should: boolean) => void
 		onAttachListener?: (client: MercuryClient) => void
 	}) {
-		MercuryClientFactory.setIsTestMode(true)
-
-		this.cwd = await this.setupSkillDir('registered-skill')
-
-		const { skill, client } = await this.Fixture('skill').loginAsDemoSkill({
-			name: 'skill1',
-		})
-
-		const eventName = `my-cool-event::v2021_01_22`
-		const fqen = `${skill.slug}.${eventName}`
-
-		await this.registerEvents(client, eventName)
-		//@ts-ignore
-		client.mixinContract(this.buildContract(fqen))
-		this.setupListenersForEventsRegisteredBySkill(skill)
-		this.generateGoodContractFileForSkill(skill)
-
-		process.env.SKILL_ID = skill.id
-		process.env.SKILL_API_KEY = skill.apiKey
-
-		if (options?.onUnregisterListeners) {
-			const client2 = await this.Fixture('mercury').connectToApi()
-			await client2.on('unregister-listeners::v2020_12_25', async () => {
-				options?.onUnregisterListeners?.()
-				return { unregisterCount: 0 }
-			})
-		}
-
-		const bootedSkill = await this.Skill()
-		const events = bootedSkill.getFeatureByCode('event') as EventFeaturePlugin
-
-		if (options?.onAttachListeners) {
-			//@ts-ignore
-			const oldAttachListeners = events.attachListeners.bind(events)
-
-			//@ts-ignore
-			events.attachListeners = async (client: any) => {
-				const results = await oldAttachListeners(client)
-				options?.onAttachListeners?.(client)
-				return results
-			}
-		}
-
-		if (options?.onSetShouldAutoRegisterListeners) {
-			const oldConnect = events.connectToApi.bind(events)
-			events.connectToApi = async (connectOptions: any) => {
-				const client = await oldConnect(connectOptions)
-
-				//@ts-ignore
-				client.setShouldAutoRegisterListeners = (should: boolean) => {
-					//@ts-ignore
-					client.shouldAutoRegisterListeners = should
-					options?.onSetShouldAutoRegisterListeners?.(should)
-				}
-
-				//@ts-ignore
-				client.on = () => {
-					//@ts-ignore
-					options?.onAttachListener?.(client)
-				}
-
-				return client
-			}
-		}
-
-		return { bootedSkill, events }
+		this.cwd = await this.generateSkillFromTestPath('registered-skill')
+		return this.EventFixture().registerSkillAndSetupListeners(options)
 	}
 }