+++ packages/spruce-skill-booter/src/skills/Skill.ts
@@ -5,6 +5,10 @@ import {
 	Skill as ISkill,
 	SkillContext,
 	SkillFeature,
+	pluginUtil,
+	LogOptions,
+	Level,
+	diskUtil,
 } from '@sprucelabs/spruce-skill-utils'
 import SpruceError from '../errors/SpruceError'
 
@@ -22,7 +26,10 @@ export default class Skill implements ISkill {
 	public hashSpruceDir
 
 	private featureMap: Record<string, SkillFeature> = {}
-	private log: Log
+	private _log: Log
+	private get log() {
+		return this._log
+	}
 	private _isRunning = false
 	private shutdownTimeout: any
 	private isKilling = false
@@ -35,7 +42,8 @@ export default class Skill implements ISkill {
 		this.rootDir = options.rootDir
 		this.activeDir = options.activeDir
 		this.hashSpruceDir = options.hashSpruceDir
-		this.log = options.log ?? buildLog('Skill')
+		this._log = options.log ?? this.buildLogWithTransports()
+
 		this.shouldCountdownOnExit = options.shouldCountdownOnExit ?? true
 	}
 
@@ -217,5 +225,56 @@ export default class Skill implements ISkill {
 		value: SkillContext[Key]
 	) {
 		this.context[key] = value
+	}
+
+	private buildLogWithTransports() {
+		let transportsByLevel:
+			| Partial<LogOptions['transportsByLevel']>
+			| undefined = undefined
+		const lookupDir = diskUtil.resolvePath(this.activeDir, 'logTransports')
+
+		if (diskUtil.doesDirExist(lookupDir)) {
+			const matches = pluginUtil.importSync([], lookupDir)
+
+			if (matches.length > 0) {
+				transportsByLevel = {}
+
+				for (const first of matches) {
+					if (!first) {
+						continue
+					}
+					if (!Array.isArray(first.levels)) {
+						throw new SpruceError({
+							//@ts-ignore
+							code: 'INVALID_LOG_TRANSPORT',
+							friendlyMessage: `The log transport you supplied is invalid. Make sure it exports a function by default and returns an object that looks like:
+							
+			{ 
+				level: string[], 
+				transport: (...messageParts: string[]) => void
+			}
+			`,
+						})
+					}
+
+					first.levels.forEach((level: Level) => {
+						if (transportsByLevel) {
+							if (transportsByLevel[level]) {
+								throw new SpruceError({
+									//@ts-ignore
+									code: 'DUPLICATE_LOG_TRANSPORT',
+									friendlyMessage: `You have two transports handling '${level}' and that is not supported.`,
+								})
+							}
+							transportsByLevel[level] = first.transport
+						}
+					})
+				}
+			}
+		}
+
+		return buildLog(`Skill (${process.pid})`, {
+			transportsByLevel,
+		})
 	}
 }