+++ packages/spruce-test-fixtures/src/__tests__/behavioral/OrganizationFixture.test.ts
@@ -1,7 +1,10 @@
 import { eventErrorAssertUtil } from '@sprucelabs/spruce-event-utils'
 import { test, assert } from '@sprucelabs/test'
-import { errorAssertUtil } from '@sprucelabs/test-utils'
 import AbstractSpruceFixtureTest from '../../tests/AbstractSpruceFixtureTest'
+import {
+	DEMO_NUMBER_HIRING,
+	DEMO_NUMBER_INSTALLING_SKILLS,
+} from '../../tests/constants'
 import OrganizationFixture from '../../tests/fixtures/OrganizationFixture'
 
 export default class OrganizationFixtureTest extends AbstractSpruceFixtureTest {
@@ -45,9 +48,7 @@ export default class OrganizationFixtureTest extends AbstractSpruceFixtureTest {
 		const people = this.Fixture('person')
 		const org = await this.fixture.seedDemoOrg({ name: 'my org' })
 
-		const { person } = await people.loginAsDemoPerson(
-			process.env.DEMO_NUMBER_HIRING
-		)
+		const { person } = await people.loginAsDemoPerson(DEMO_NUMBER_HIRING)
 
 		const isHired = await this.fixture.isPartOfOrg(person.id, org.id)
 		assert.isFalse(isHired)
@@ -58,9 +59,7 @@ export default class OrganizationFixtureTest extends AbstractSpruceFixtureTest {
 		const people = this.Fixture('person')
 		const org = await this.fixture.seedDemoOrg({ name: 'my org' })
 
-		const { person } = await people.loginAsDemoPerson(
-			process.env.DEMO_NUMBER_HIRING
-		)
+		const { person } = await people.loginAsDemoPerson(DEMO_NUMBER_HIRING)
 
 		await this.fixture.addPerson({
 			personId: person.id,
@@ -74,42 +73,60 @@ export default class OrganizationFixtureTest extends AbstractSpruceFixtureTest {
 
 	@test()
 	protected static async isNotInstalledByDefault() {
-		const { skill, org } = await OrganizationFixtureTest.seedOrgAndSkill()
+		const { skill, org } = await this.seedOrgAndSkill()
 		const isInstalled = await this.fixture.isSkillInstalled(skill.id, org.id)
 		assert.isFalse(isInstalled)
 	}
 
 	@test()
 	protected static async showsAsInstalled() {
-		const { skill, org } = await OrganizationFixtureTest.seedOrgAndSkill()
+		const { skill, org } = await this.seedOrgAndSkill()
 		await this.fixture.installSkill(skill.id, org.id)
 		await this.assertSkillIsInstalled(skill.id, org.id)
 	}
 
 	@test()
 	protected static async cantInstallWithBadSkill() {
-		const { org } = await OrganizationFixtureTest.seedOrgAndSkill()
+		const { org } = await this.seedOrgAndSkill()
 		const err = await assert.doesThrowAsync(() =>
-			this.fixture.installSkillsBySlug({
+			this.fixture.installSkillsByNamespace({
 				organizationId: org.id,
-				slugs: ['aoeuaoeu'],
+				namespaces: ['aoeuaoeu'],
 			})
 		)
 
-		errorAssertUtil.assertError(err, 'SKILL_NOT_FOUND', {
-			slug: 'aoeuaoeu',
+		eventErrorAssertUtil.assertError(err, 'INVALID_NAMESPACES', {
+			namespaces: ['aoeuaoeu'],
 		})
 	}
 
 	@test()
 	protected static async canInstallWithSlug() {
-		const { skill, org } = await OrganizationFixtureTest.seedOrgAndSkill()
-		await this.fixture.installSkillsBySlug({
+		const { skill, org } = await this.seedOrgAndSkill()
+		await this.fixture.installSkillsByNamespace({
 			organizationId: org.id,
-			slugs: [skill.slug],
+			namespaces: [skill.slug],
 		})
 
 		await this.assertSkillIsInstalled(skill.id, org.id)
+	}
+
+	@test()
+	protected static async canInstallSkillFromAnotherPerson() {
+		const { skill, org } = await this.seedOrgAndSkill()
+
+		const skill2 = await this.Fixture('skill').seedDemoSkill({
+			name: 'testing testy 2',
+			creatorPhone: DEMO_NUMBER_INSTALLING_SKILLS,
+		})
+
+		await this.fixture.installSkillsByNamespace({
+			organizationId: org.id,
+			namespaces: [skill.slug, skill2.slug],
+		})
+
+		await this.assertSkillIsInstalled(skill.id, org.id)
+		await this.assertSkillIsInstalled(skill2.id, org.id)
 	}
 
 	private static async assertSkillIsInstalled(skillId: string, orgId: string) {