+++ packages/spruce-test-fixtures/src/__tests__/behavioral/StoreFixture.test.ts
@@ -1,6 +1,11 @@
-import { StoreLoader } from '@sprucelabs/data-stores'
+import {
+	DatabaseFixture,
+	StoreLoader,
+	StoreFactory,
+} from '@sprucelabs/data-stores'
 import AbstractSpruceTest, { assert, test } from '@sprucelabs/test'
 import FixtureFactory from '../../tests/fixtures/FixtureFactory'
+import StoreFixture from '../../tests/fixtures/StoreFixture'
 
 declare module '@sprucelabs/data-stores/build/types/stores.types' {
 	interface StoreMap {
@@ -9,13 +14,19 @@ declare module '@sprucelabs/data-stores/build/types/stores.types' {
 }
 
 export default class StoreFixtureTest extends AbstractSpruceTest {
+	private static originalDestroy: () => Promise<void>
+	protected static async beforeAll() {
+		await super.beforeAll()
+		this.originalDestroy = DatabaseFixture.destroy.bind(DatabaseFixture)
+	}
 	protected static async beforeEach() {
 		await super.beforeEach()
+		DatabaseFixture.destroy = this.originalDestroy
 	}
 
 	@test()
 	protected static async canGetFixture() {
-		assert.isTruthy(this.Fixture().Fixture('store'))
+		assert.isTruthy(this.FixtureFactory().Fixture('store'))
 	}
 
 	@test()
@@ -30,17 +41,69 @@ export default class StoreFixtureTest extends AbstractSpruceTest {
 			)
 		)
 
-		const dbFixture = this.Fixture().Fixture('database')
+		const dbFixture = this.FixtureFactory().Fixture('database')
 		const db = await dbFixture.connectToDatabase()
 
 		StoreLoader.setDatabase(db)
 
-		const fixture = this.Fixture().Fixture('store')
+		const fixture = this.FixtureFactory().Fixture('store')
 		const goodStore = await fixture.Store('good')
 		assert.isTruthy(goodStore)
 	}
 
-	private static Fixture() {
+	@test()
+	protected static async beforeAllCallsBeforeAllOnDatabaseFixture() {
+		let wasHit = false
+
+		//@ts-ignore
+		DatabaseFixture.beforeAll = () => {
+			wasHit = true
+		}
+
+		await StoreFixture.beforeAll()
+
+		assert.isTrue(wasHit)
+	}
+
+	@test()
+	protected static async beforeEachCallsDestroyOnDatabaseFixture() {
+		let wasHit = false
+
+		//@ts-ignore
+		DatabaseFixture.destroy = () => {
+			wasHit = true
+		}
+
+		await StoreFixture.beforeEach()
+
+		assert.isTrue(wasHit)
+	}
+
+	@test()
+	protected static async storeLoaderUsesNewInstanceOfDbEachRun() {
+		const loader1 = await StoreLoader.getInstance()
+
+		await StoreFixture.beforeEach()
+
+		const loader2 = await StoreLoader.getInstance()
+
+		//@ts-ignore
+		assert.isNotEqual(loader1.db, loader2.db)
+	}
+
+	@test()
+	protected static async storeFixtureCallsResetOnStoreFactory() {
+		let wasHit = false
+		//@ts-ignore
+		StoreFactory.reset = () => {
+			wasHit = true
+		}
+
+		await StoreFixture.beforeEach()
+		assert.isTrue(wasHit)
+	}
+
+	private static FixtureFactory() {
 		return new FixtureFactory({ cwd: this.cwd })
 	}
 }