+++ packages/spruce-skill-booter/CHANGELOG.md
@@ -3,6 +3,225 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [5.4.23](https://github.com/sprucelabsai/spruce-features-workspace/compare/v5.4.22...v5.4.23) (2021-08-21)
+
+**Note:** Version bump only for package @sprucelabs/spruce-skill-booter
+
+
+
+
+
+## [5.4.22](https://github.com/sprucelabsai/spruce-features-workspace/compare/v5.4.21...v5.4.22) (2021-08-20)
+
+**Note:** Version bump only for package @sprucelabs/spruce-skill-booter
+
+
+
+
+
+## [5.4.21](https://github.com/sprucelabsai/spruce-features-workspace/compare/v5.4.20...v5.4.21) (2021-08-19)
+
+**Note:** Version bump only for package @sprucelabs/spruce-skill-booter
+
+
+
+
+
+## [5.4.20](https://github.com/sprucelabsai/spruce-features-workspace/compare/v5.4.19...v5.4.20) (2021-08-18)
+
+**Note:** Version bump only for package @sprucelabs/spruce-skill-booter
+
+
+
+
+
+## [5.4.19](https://github.com/sprucelabsai/spruce-features-workspace/compare/v5.4.18...v5.4.19) (2021-08-18)
+
+**Note:** Version bump only for package @sprucelabs/spruce-skill-booter
+
+
+
+
+
+## [5.4.18](https://github.com/sprucelabsai/spruce-features-workspace/compare/v5.4.17...v5.4.18) (2021-08-17)
+
+**Note:** Version bump only for package @sprucelabs/spruce-skill-booter
+
+
+
+
+
+## [5.4.17](https://github.com/sprucelabsai/spruce-features-workspace/compare/v5.4.16...v5.4.17) (2021-08-16)
+
+**Note:** Version bump only for package @sprucelabs/spruce-skill-booter
+
+
+
+
+
+## [5.4.16](https://github.com/sprucelabsai/spruce-features-workspace/compare/v5.4.15...v5.4.16) (2021-08-16)
+
+**Note:** Version bump only for package @sprucelabs/spruce-skill-booter
+
+
+
+
+
+## [5.4.15](https://github.com/sprucelabsai/spruce-features-workspace/compare/v5.4.14...v5.4.15) (2021-08-14)
+
+**Note:** Version bump only for package @sprucelabs/spruce-skill-booter
+
+
+
+
+
+## [5.4.14](https://github.com/sprucelabsai/spruce-features-workspace/compare/v5.4.13...v5.4.14) (2021-08-13)
+
+**Note:** Version bump only for package @sprucelabs/spruce-skill-booter
+
+
+
+
+
+## [5.4.13](https://github.com/sprucelabsai/spruce-features-workspace/compare/v5.4.12...v5.4.13) (2021-08-12)
+
+**Note:** Version bump only for package @sprucelabs/spruce-skill-booter
+
+
+
+
+
+## [5.4.12](https://github.com/sprucelabsai/spruce-features-workspace/compare/v5.4.11...v5.4.12) (2021-08-12)
+
+**Note:** Version bump only for package @sprucelabs/spruce-skill-booter
+
+
+
+
+
+## [5.4.11](https://github.com/sprucelabsai/spruce-features-workspace/compare/v5.4.10...v5.4.11) (2021-08-11)
+
+**Note:** Version bump only for package @sprucelabs/spruce-skill-booter
+
+
+
+
+
+## [5.4.10](https://github.com/sprucelabsai/spruce-features-workspace/compare/v5.4.9...v5.4.10) (2021-08-11)
+
+**Note:** Version bump only for package @sprucelabs/spruce-skill-booter
+
+
+
+
+
+## [5.4.9](https://github.com/sprucelabsai/spruce-features-workspace/compare/v5.4.8...v5.4.9) (2021-08-10)
+
+**Note:** Version bump only for package @sprucelabs/spruce-skill-booter
+
+
+
+
+
+## [5.4.8](https://github.com/sprucelabsai/spruce-features-workspace/compare/v5.4.7...v5.4.8) (2021-08-10)
+
+**Note:** Version bump only for package @sprucelabs/spruce-skill-booter
+
+
+
+
+
+## [5.4.7](https://github.com/sprucelabsai/spruce-features-workspace/compare/v5.4.6...v5.4.7) (2021-08-09)
+
+**Note:** Version bump only for package @sprucelabs/spruce-skill-booter
+
+
+
+
+
+## [5.4.6](https://github.com/sprucelabsai/spruce-features-workspace/compare/v5.4.5...v5.4.6) (2021-08-09)
+
+**Note:** Version bump only for package @sprucelabs/spruce-skill-booter
+
+
+
+
+
+## [5.4.5](https://github.com/sprucelabsai/spruce-features-workspace/compare/v5.4.4...v5.4.5) (2021-08-09)
+
+**Note:** Version bump only for package @sprucelabs/spruce-skill-booter
+
+
+
+
+
+## [5.4.4](https://github.com/sprucelabsai/spruce-features-workspace/compare/v5.4.3...v5.4.4) (2021-08-09)
+
+**Note:** Version bump only for package @sprucelabs/spruce-skill-booter
+
+
+
+
+
+## [5.4.3](https://github.com/sprucelabsai/spruce-features-workspace/compare/v5.4.2...v5.4.3) (2021-08-08)
+
+**Note:** Version bump only for package @sprucelabs/spruce-skill-booter
+
+
+
+
+
+## [5.4.2](https://github.com/sprucelabsai/spruce-features-workspace/compare/v5.4.1...v5.4.2) (2021-08-08)
+
+**Note:** Version bump only for package @sprucelabs/spruce-skill-booter
+
+
+
+
+
+## [5.4.1](https://github.com/sprucelabsai/spruce-features-workspace/compare/v5.4.0...v5.4.1) (2021-08-06)
+
+**Note:** Version bump only for package @sprucelabs/spruce-skill-booter
+
+
+
+
+
+# [5.4.0](https://github.com/sprucelabsai/spruce-features-workspace/compare/v5.3.2...v5.4.0) (2021-08-06)
+
+
+### Features
+
+* log transport loading on skill boot ([5c17094](https://github.com/sprucelabsai/spruce-features-workspace/commit/5c17094))
+
+
+
+
+
+## [5.3.2](https://github.com/sprucelabsai/spruce-features-workspace/compare/v5.3.1...v5.3.2) (2021-08-05)
+
+**Note:** Version bump only for package @sprucelabs/spruce-skill-booter
+
+
+
+
+
+## [5.3.1](https://github.com/sprucelabsai/spruce-features-workspace/compare/v5.3.0...v5.3.1) (2021-08-05)
+
+**Note:** Version bump only for package @sprucelabs/spruce-skill-booter
+
+
+
+
+
+# [5.3.0](https://github.com/sprucelabsai/spruce-features-workspace/compare/v5.2.3...v5.3.0) (2021-08-04)
+
+**Note:** Version bump only for package @sprucelabs/spruce-skill-booter
+
+
+
+
+
 ## [5.2.3](https://github.com/sprucelabsai/spruce-features-workspace/compare/v5.2.2...v5.2.3) (2021-08-03)
 
 **Note:** Version bump only for package @sprucelabs/spruce-skill-booter